% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_who_risk.R
\name{country_to_region}
\alias{country_to_region}
\title{Map Country Code to Region}
\usage{
country_to_region(country)
}
\arguments{
\item{country}{Character vector of ISO 3166-1 alpha-3 country codes.}
}
\value{
Character vector of region names.
}
\description{
Utility function to map ISO 3166-1 alpha-3 country codes to WHO/GBD regions.
}
\examples{
country_to_region(c("USA", "DEU", "JPN"))

}
